Contracts und Datentyp angepasst
Contracts hinzugefügt und Datentypen der Datenbank angepasst und durch optionale Parameter die Aufrufe editiert damit diese wieder funktionieren.

+++ Lagerverwaltung/Views/Produktliste.cs
@@ -22,10 +22,10 @@ namespace Lagerverwaltung.Views
             _produktListe = new List<Core.Produkt>();
 
             //TEST
-            _produktListe.Add(new Core.Produkt(0, "Eimer", "31M32", (float)40.2, (float)40.3, 450403, 923425, new Core.Artikeltyp(0, "bla", 50, 60), 40, 60, 80));
-            _produktListe.Add(new Core.Produkt(1, "Tisch", "T15-CH", (float)100.23, (float)1.01, 4567, 3425, new Core.Artikeltyp(1, "blubb", 50, 60), 1000, 2000, 500));
-            _produktListe.Add(new Core.Produkt(2, "Dildo", "D-1L-D0", (float)1.15, (float)9000, 12345, 54321, new Core.Artikeltyp(0, "bla", 50, 60), 300, 100, 100));
-            _produktListe.Add(new Core.Produkt(3, "Thermoskanne voll Mett", "N0-T4-5CH-3N-Mu-Mu", (float)500, (float)1500, 1235634, 23515, new Core.Artikeltyp(0, "bla", 50, 60), 500, 500, 500));
+            _produktListe.Add(new Core.Produkt(0, "Eimer", "31M32", 40.2f, 40.3f, 450403, 923425, new Core.Artikeltyp(0, "bla", 50, 60), 40, 60, 80));
+            _produktListe.Add(new Core.Produkt(1, "Tisch", "T15-CH", 100.23f, 1.01f, 4567, 3425, new Core.Artikeltyp(1, "blubb", 50, 60), 1000, 2000, 500));
+            _produktListe.Add(new Core.Produkt(2, "Dildo", "D-1L-D0", 1.15f, 9000, 12345, 54321, new Core.Artikeltyp(0, "bla", 50, 60), 300, 100, 100));
+            _produktListe.Add(new Core.Produkt(3, "Thermoskanne voll Mett", "N0-T4-5CH-3N-Mu-Mu", 500, 1500, 1235634, 23515, new Core.Artikeltyp(0, "bla", 50, 60), 500, 500, 500));
             //TEST
 
             foreach (Core.Produkt produkt in _produktListe)
